initial merge with cli that compiles
