+++ peernode-market/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [[bin]] # Bin to run the gRPC client
 name = "peer-node"
-path = "src/peer-node/client.rs"
+path = "src/client.rs"
 
 [dependencies]
 tonic = "0.11"
