+++ peernode-market/config.json
@@ -1,0 +1,1 @@
+{"name":"default","files":{},"prices":{},"tokens":{},"port":"8080","bootstrap_peers":[],"listen_addr":"/ip4/0.0.0.0/tcp/6881","private_key":"private_key.pk8"}
